ARC merging, bug fixes and design changes

+++ SONIC THE HEDGEHOG (2006) Mod Manager/ModManager.cs
@@ -10,7 +10,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace SONIC_THE_HEDGEHOG__2006__Mod_Manager
+namespace Sonic_06_Mod_Manager
 {
     public partial class ModManager : Form
     {
@@ -34,9 +34,81 @@ namespace SONIC_THE_HEDGEHOG__2006__Mod_Manager
             xeniaBox.Text = Properties.Settings.Default.xeniaPath;
         }
 
+        public string applicationData = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
+
         private void ModManager_Load(object sender, EventArgs e)
         {
+            if (!Directory.Exists($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool")) Directory.CreateDirectory($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool");
+            if (!File.Exists($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool\\arctool.php")) File.WriteAllBytes($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool\\arctool.php", Properties.Resources.arctoolphp);
+            if (!File.Exists($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool\\unarc.php")) File.WriteAllBytes($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool\\unarc.php", Properties.Resources.unarcphp);
+            if (!File.Exists($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool\\arcc.php")) File.WriteAllBytes($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool\\arcc.php", Properties.Resources.arccphp);
+            if (!File.Exists($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool.exe")) File.WriteAllBytes($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool.exe", Properties.Resources.arctool);
+
             RefreshMods();
+            CleanUpMods();
+        }
+
+        private void MergeARCs(string arc1, string arc2, string output)
+        {
+            string tempPath = $"{applicationData}\\Temp\\{Path.GetRandomFileName()}"; Directory.CreateDirectory(tempPath);
+            var tempData = new DirectoryInfo(tempPath);
+            File.Copy(arc1, Path.Combine(tempPath, Path.GetFileName(arc1)));
+            File.Move(origArcPath, targetArcPath);
+            ProcessStartInfo arctool;
+
+            arctool = new ProcessStartInfo($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool.exe", $"-d \"{Path.Combine(tempPath, Path.GetFileName(arc1))}\"")
+            {
+                WorkingDirectory = $"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\",
+                WindowStyle = ProcessWindowStyle.Hidden
+            };
+
+            var Unpack1 = Process.Start(arctool);
+            Unpack1.WaitForExit();
+            Unpack1.Close();
+
+            File.Delete(Path.Combine(tempPath, Path.GetFileName(arc1)));
+
+            //if (arc1 != arc2) { Directory.Move(Path.Combine(tempPath, Path.GetFileNameWithoutExtension(arc1)), Path.Combine(tempPath, Path.GetFileNameWithoutExtension(arc2))); }
+
+            File.Copy(arc2, Path.Combine(tempPath, Path.GetFileName(arc2)));
+
+            arctool = new ProcessStartInfo($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool.exe", $"-d \"{Path.Combine(tempPath, Path.GetFileName(arc2))}\"")
+            {
+                WorkingDirectory = $"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\",
+                WindowStyle = ProcessWindowStyle.Hidden
+            };
+
+            var Unpack2 = Process.Start(arctool);
+            Unpack2.WaitForExit();
+            Unpack2.Close();
+
+            File.Delete(Path.Combine(tempPath, Path.GetFileName(arc2)));
+
+            arctool = new ProcessStartInfo($"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\arctool.exe", $"-i \"{Path.Combine(tempPath, Path.GetFileNameWithoutExtension(arc2))}\" -c \"{output}\"")
+            {
+                WorkingDirectory = $"{applicationData}\\Sonic_06_Mod_Manager\\Tools\\",
+                WindowStyle = ProcessWindowStyle.Hidden
+            };
+
+            var Repack1 = Process.Start(arctool);
+            Repack1.WaitForExit();
+            Repack1.Close();
+
+            try
+            {
+                if (Directory.Exists(tempPath))
+                {
+                    foreach (FileInfo file in tempData.GetFiles())
+                    {
+                        file.Delete();
+                    }
+                    foreach (DirectoryInfo directory in tempData.GetDirectories())
+                    {
+                        directory.Delete(true);
+                    }
+                }
+            }
+            catch { return; }
         }
 
         private void RefreshMods()
@@ -58,25 +130,25 @@ namespace SONIC_THE_HEDGEHOG__2006__Mod_Manager
                 {
                     modsPath = modPathBrowser.SelectedPath;
                     modsBox.Text = modsPath;
-
-                    if (!Directory.Exists(modsPath))
-                    {
-                        modList.Items.Clear();
-                        return;
-                    }
-                    modArray = Directory.GetDirectories(modsPath);
-                    modList.Items.Clear();
-                    foreach (string mod in modArray)
-                    {
-                        var modName = mod.Remove(0, Path.GetDirectoryName(mod).Length);
-                        modName = modName.Replace("\\", "");
-                        modList.Items.Add(modName);
-                    }
-                    Properties.Settings.Default.modsPath = modsPath;
-                    Properties.Settings.Default.Save();
                 }
                 else{ Application.Exit(); }
             }
+
+            if (!Directory.Exists(modsPath))
+            {
+                modList.Items.Clear();
+                return;
+            }
+            modArray = Directory.GetDirectories(modsPath);
+            modList.Items.Clear();
+            foreach (string mod in modArray)
+            {
+                var modName = mod.Remove(0, Path.GetDirectoryName(mod).Length);
+                modName = modName.Replace("\\", "");
+                modList.Items.Add(modName);
+            }
+            Properties.Settings.Default.modsPath = modsPath;
+            Properties.Settings.Default.Save();
         }
 
         private void RefreshButton_Click(object sender, EventArgs e)
@@ -101,8 +173,8 @@ namespace SONIC_THE_HEDGEHOG__2006__Mod_Manager
         private void XeniaButton_Click(object sender, EventArgs e)
         {
             OpenFileDialog xeniaBrowser = new OpenFileDialog();
-            xeniaBrowser.Title = "Select Xenia Executable";
-            xeniaBrowser.Filter = "Xenia Emulator (*.exe)|*.exe";
+            xeniaBrowser.Title = "Select a Xenia executable...";
+            xeniaBrowser.Filter = "Programs (*.exe)|*.exe";
             xeniaBrowser.FilterIndex = 1;
             xeniaBrowser.RestoreDirectory = true;
             if (xeniaBrowser.ShowDialog() == DialogResult.OK)
@@ -140,9 +212,17 @@ namespace SONIC_THE_HEDGEHOG__2006__Mod_Manager
             if (modsBox.Text != string.Empty && s06PathBox.Text != string.Empty && xeniaBox.Text != string.Empty)
             {
                 CleanUpMods();
-                CopyMods();
+                try { CopyMods(); }
+                catch (Exception ex)
+                {
+                    MessageBox.Show($"Please refer to the following error for more information:\n\n{ex}", "Fatal Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    CleanUpMods();
+                    return;
+                }
                 Console.WriteLine("\nStarting Xenia.\n");
-                var xenia = Process.Start(xeniaPath);
+                ProcessStartInfo xeniaExec;
+                xeniaExec = new ProcessStartInfo(xeniaPath) { WorkingDirectory = Path.GetDirectoryName(xeniaPath) };
+                var xenia = Process.Start(xeniaExec);
                 xenia.WaitForExit();
                 CleanUpMods();
             }
@@ -151,17 +231,20 @@ namespace SONIC_THE_HEDGEHOG__2006__Mod_Manager
 
         private void CleanUpMods()
         {
-            var mods = Directory.GetFiles(s06Path, "*.*", SearchOption.AllDirectories)
-            .Where(s => s.EndsWith(".arc_back") || s.EndsWith(".wmv_back") || s.EndsWith(".xma_back"));
-
-            foreach (var mod in mods)
+            if (s06Path != string.Empty)
             {
-                if (File.Exists(mod.ToString().Remove(mod.Length - 5)))
+                var mods = Directory.GetFiles(s06Path, "*.*", SearchOption.AllDirectories)
+                .Where(s => s.EndsWith(".arc_back") || s.EndsWith(".wmv_back") || s.EndsWith(".xma_back") || s.EndsWith(".xex_back"));
+
+                foreach (var mod in mods)
                 {
-                    File.Delete(mod.ToString().Remove(mod.Length - 5));
+                    if (File.Exists(mod.ToString().Remove(mod.Length - 5)))
+                    {
+                        File.Delete(mod.ToString().Remove(mod.Length - 5));
+                    }
+                    File.Move(mod.ToString(), mod.ToString().Remove(mod.Length - 5));
+                    Console.WriteLine("Removing: " + mod.ToString().Remove(mod.Length - 5));
                 }
-                File.Move(mod.ToString(), mod.ToString().Remove(mod.Length - 5));
-                Console.WriteLine("Removing: " + mod.ToString().Remove(mod.Length - 5));
             }
         }
 
@@ -171,25 +254,42 @@ namespace SONIC_THE_HEDGEHOG__2006__Mod_Manager
             foreach (var item in indexes)
             {
                 var mods = Directory.GetFiles(modArray[item], "*.*", SearchOption.AllDirectories)
-                .Where(s => s.EndsWith(".arc") || s.EndsWith(".wmv") || s.EndsWith(".xma"));
+                .Where(s => s.EndsWith(".arc") || s.EndsWith(".wmv") || s.EndsWith(".xma") || s.EndsWith(".xex"));
 
                 foreach (var mod in mods)
                 {
                     arcPath = mod.Remove(0, modArray[item].Length);
                     var arcName = mod.Remove(0, Path.GetDirectoryName(mod).Length);
                     arcName = arcName.Replace("\\", "");
-                    Console.WriteLine("Copying " + mod);
                     origArcPath = s06Path + arcPath;
                     targetArcPath = origArcPath + "_back";
 
-                    if (!File.Exists(targetArcPath))
+                    if (File.ReadAllLines(modArray[modList.SelectedIndex] + "\\mod.ini").Contains("Merge=\"True\""))
                     {
-                        File.Move(origArcPath, targetArcPath);
-                        File.Copy(mod, origArcPath);
+                        if (mod.EndsWith(".arc"))
+                        {
+                            Console.WriteLine("Merging " + mod);
+                            MergeARCs(origArcPath, mod, origArcPath);
+                        }
+                        else
+                        {
+                            Console.WriteLine("Copying " + mod);
+                            File.Move(origArcPath, targetArcPath);
+                            File.Copy(mod, origArcPath);
+                        }
                     }
                     else
                     {
-                        Console.WriteLine("Skipped " + mod);
+                        if (!File.Exists(targetArcPath))
+                        {
+                            Console.WriteLine("Copying " + mod);
+                            File.Move(origArcPath, targetArcPath);
+                            File.Copy(mod, origArcPath);
+                        }
+                        else
+                        {
+                            Console.WriteLine("Skipped " + mod);
+                        }
                     }
                 }
             }
@@ -216,29 +316,35 @@ namespace SONIC_THE_HEDGEHOG__2006__Mod_Manager
                             {
                                 entryValue = line.Substring(line.IndexOf("=") + 2);
                                 entryValue = entryValue.Remove(entryValue.Length - 1);
-                                modDetails = modDetails + "Mod Title: " + entryValue + "\n";
+                                modDetails = modDetails + "Title: " + entryValue + "\n";
                             }
                             if (line.Contains("Version=\""))
                             {
                                 entryValue = line.Substring(line.IndexOf("=") + 2);
                                 entryValue = entryValue.Remove(entryValue.Length - 1);
-                                modDetails = modDetails + "Mod Version Number: " + entryValue + "\n";
+                                modDetails = modDetails + "Version: " + entryValue + "\n";
                             }
                             if (line.Contains("Date=\""))
                             {
                                 entryValue = line.Substring(line.IndexOf("=") + 2);
                                 entryValue = entryValue.Remove(entryValue.Length - 1);
-                                modDetails = modDetails + "Mod Creation Date: " + entryValue + "\n";
+                                modDetails = modDetails + "Date: " + entryValue + "\n";
                             }
                             if (line.Contains("Author=\""))
                             {
                                 entryValue = line.Substring(line.IndexOf("=") + 2);
                                 entryValue = entryValue.Remove(entryValue.Length - 1);
-                                modDetails = modDetails + "Mod Author: " + entryValue + "\n";
+                                modDetails = modDetails + "Author: " + entryValue + "\n";
+                            }
+                            if (line.Contains("Merge=\""))
+                            {
+                                entryValue = line.Substring(line.IndexOf("=") + 2);
+                                entryValue = entryValue.Remove(entryValue.Length - 1);
+                                modDetails = modDetails + "Merge: " + entryValue + "\n";
                             }
                         }
                     }
-                    MessageBox.Show(modDetails, "Mod Information for " + modList.Items[modList.SelectedIndex]);
+                    MessageBox.Show(modDetails, modList.Items[modList.SelectedIndex].ToString(), MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
                 else
                 {
@@ -267,7 +373,12 @@ namespace SONIC_THE_HEDGEHOG__2006__Mod_Manager
 
         private void Button2_Click(object sender, EventArgs e)
         {
-            MessageBox.Show("Sonic '06 Mod Manager\n\nKnuxfan24 - Lead Developer\nHyper - Project Helper",
+            MessageBox.Show("Sonic '06 Mod Manager\n\n" +
+                            "" +
+                            "Knuxfan24 - Lead Developer\n" +
+                            "Hyper - Co-developer\n" +
+                            "xose - ARC Unpacker\n" +
+                            "g0ldenlink - ARC Repacker",
                             "Credits",
                             MessageBoxButtons.OK,
                             MessageBoxIcon.Information);
